+++ patches/Accessibility/strings_it.cpp
@@ -450,7 +450,7 @@ const char* Get(Id id) {
         case Id::ModSettingExtendedCycling:   return "Selezione oggetti su tutta la mappa";
         case Id::ModSettingRoomShapes:        return "Descrizioni della forma delle stanze";
         case Id::ModSettingWallSounds:        return "Suoni dei muri";
-        case Id::ModSettingHumanSubtitles:    return "Leggi i sottotitoli dei parlanti umani";
+        case Id::ModSettingHumanSubtitles:    return "Leggi i sottotitoli dei parlanti doppiati";
         case Id::ModSettingTurretAutoAim:     return "Mira automatica";
         case Id::ModSettingSkipIntros:        return "Salta i video introduttivi";
         case Id::ModSettingSkipIntrosOnNextLaunch: return "I video introduttivi verranno saltati al prossimo avvio.";